requirements.txt
attrs>=16.2